requirements.txt
google-genai>=1.0.0
twelvelabs>=0.4.0
boto3>=1.35.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.66.0
tabulate>=0.9.0
decord>=0.6.0
av>=12.0.0
opencv-python>=4.8.0
